deepvape-main/backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Werkzeug==2.3.7
gunicorn==21.2.0
psycopg2-binary==2.9.7
PyMySQL==1.1.0
python-dotenv==1.0.0
Pillow>=10.0.0
python-telegram-bot==20.7 